moved new workspace button in headers and CreateWorkspaceModal component to open create workspace form from any menu not just in Workspace menu
